refactor(result): one output-format serializer per file
Pure structural move (no behavior change): scanning/result.rs (1249 lines)
is now scanning/result/ —
- mod.rs            the model (FindingType/Method/Confidence, Result,
                    ResultBuilder, Sanitized/ScanMetadata) + the shared
                    to_sanitized / to_json_value / make_scan_meta_value core
- format_json.rs    results_to_json / _jsonl
- format_toml.rs    results_to_toml(_with_meta)
- format_markdown.rs results_to_markdown(_with_meta)  (~250 lines)
- format_sarif.rs   results_to_sarif(_with_meta) + sarif_rule_id/for_id

Each format is its own `impl Result` block, so extending one format no
longer collides with the others. All serializers stay `pub(crate)` on
Result, so `Result::results_to_markdown_with_meta` etc. resolve unchanged;
tests.rs untouched. 28 result unit tests + 6 markdown + 12 sarif integration
tests pass, clippy clean.
