+++ src/scanning/result/mod.rs
@@ -660,590 +660,12 @@ impl Result {
         }
         value
     }
-
-    /// Serialize a slice of Result into JSON array string. Set pretty=true for pretty-printed JSON.
-    // The plain (meta-less) arms of the format matrix. `output.rs` routes every
-    // scan through the `*_with_meta` variants, so these are kept as the
-    // complete set rather than because a caller reaches them today.
-    #[allow(dead_code)]
-    pub(crate) fn results_to_json(
-        results: &[Result],
-        include_request: bool,
-        include_response: bool,
-        pretty: bool,
-    ) -> String {
-        let vals: Vec<serde_json::Value> = results
-            .iter()
-            .map(|r| r.to_json_value(include_request, include_response))
-            .collect();
-        if pretty {
-            serde_json::to_string_pretty(&vals).unwrap_or_else(|_| "[]".to_string())
-        } else {
-            serde_json::to_string(&vals).unwrap_or_else(|_| "[]".to_string())
-        }
-    }
-
-    /// Serialize a slice of Result into JSON Lines (JSONL) string.
-    #[allow(dead_code)]
-    pub(crate) fn results_to_jsonl(
-        results: &[Result],
-        include_request: bool,
-        include_response: bool,
-    ) -> String {
-        let mut out = String::new();
-        for r in results {
-            let v = r.to_json_value(include_request, include_response);
-            if let Ok(s) = serde_json::to_string(&v) {
-                out.push_str(&s);
-                out.push('\n');
-            }
-        }
-        out
-    }
-
-    /// Serialize a slice of Result into TOML string.
-    ///
-    /// For backward compatibility (public API surface under `dalfox::scanning::result`),
-    /// the 3-argument form omits the scan metadata envelope (equivalent to `meta=None`).
-    /// Use the `_with_meta` variant to carry `ScanMetadata` (targets, duration, WAF in
-    /// `target_summary`, etc.) for parity with the JSON/JSONL render path.
-    #[allow(dead_code)]
-    pub(crate) fn results_to_toml(
-        results: &[Result],
-        include_request: bool,
-        include_response: bool,
-    ) -> String {
-        Self::results_to_toml_with_meta(results, include_request, include_response, None)
-    }
-
-    /// Serialize ... with optional scan metadata (see `results_to_toml`).
-    pub(crate) fn results_to_toml_with_meta(
-        results: &[Result],
-        include_request: bool,
-        include_response: bool,
-        meta: Option<&ScanMetadata>,
-    ) -> String {
-        #[derive(Serialize)]
-        struct TomlWrapper {
-            #[serde(skip_serializing_if = "Option::is_none")]
-            meta: Option<serde_json::Value>,
-            results: Vec<SanitizedResult>,
-        }
-
-        let sanitized: Vec<SanitizedResult> = results
-            .iter()
-            .map(|r| r.to_sanitized(include_request, include_response))
-            .collect();
-
-        let meta_val = meta.map(Self::make_scan_meta_value);
-        let wrapper = TomlWrapper {
-            meta: meta_val,
-            results: sanitized,
-        };
-        toml::to_string(&wrapper).unwrap_or_else(|_| "".to_string())
-    }
-
-    /// Serialize a slice of Result into Markdown string.
-    ///
-    /// For backward compatibility (public API surface under `dalfox::scanning::result`),
-    /// the 3-argument form omits the scan metadata envelope (equivalent to `meta=None`).
-    /// Use the `_with_meta` variant to include `## Scan Metadata` + target summary tables.
-    pub fn results_to_markdown(
-        results: &[Result],
-        include_request: bool,
-        include_response: bool,
-    ) -> String {
-        Self::results_to_markdown_with_meta(results, include_request, include_response, None)
-    }
-
-    /// Serialize ... with optional scan metadata (see `results_to_markdown`).
-    pub(crate) fn results_to_markdown_with_meta(
-        results: &[Result],
-        include_request: bool,
-        include_response: bool,
-        meta: Option<&ScanMetadata>,
-    ) -> String {
-        use std::fmt::Write;
-        let mut out = String::with_capacity(results.len() * 512 + 256);
-
-        // Add header
-        out.push_str("# Dalfox Scan Results\n\n");
-
-        // Inject scan metadata envelope when provided (for parity with JSON/JSONL)
-        if let Some(m) = meta {
-            out.push_str("## Scan Metadata\n\n");
-            out.push_str("| Field | Value |\n");
-            out.push_str("|-------|-------|\n");
-            let _ = writeln!(out, "| **Dalfox Version** | {} |", m.dalfox_version);
-            let _ = writeln!(
-                out,
-                "| **Targets** | {} |",
-                m.targets.join(", ").replace('|', "\\|")
-            );
-            let _ = writeln!(out, "| **Scan Duration** | {} ms |", m.scan_duration_ms);
-            let _ = writeln!(out, "| **Total Requests** | {} |", m.total_requests);
-            if m.failed_requests > 0 {
-                let _ = writeln!(out, "| **Failed Requests** | {} |", m.failed_requests);
-            }
-            let _ = writeln!(out, "| **Findings Count** | {} |", m.findings_count);
-            // Only shown when dedup actually dropped something: a collapsed
-            // input list must be visible in the report, but the common
-            // "nothing collapsed" case doesn't need a row.
-            if m.targets_deduplicated > 0 {
-                let _ = writeln!(
-                    out,
-                    "| **Targets Deduplicated** | {} ({} mode) |",
-                    m.targets_deduplicated, m.dedup_mode
-                );
-            }
-            if let Some(b) = &m.baseline {
-                let cell = if let Some(w) = b.get("warning").and_then(|v| v.as_str()) {
-                    format!("disabled — {}", w)
-                } else {
-                    format!(
-                        "{} (mode: {}, new: {}, known: {})",
-                        b.get("path").and_then(|v| v.as_str()).unwrap_or("?"),
-                        b.get("mode").and_then(|v| v.as_str()).unwrap_or("?"),
-                        b.get("new").and_then(|v| v.as_u64()).unwrap_or(0),
-                        b.get("known").and_then(|v| v.as_u64()).unwrap_or(0),
-                    )
-                };
-                let _ = writeln!(out, "| **Baseline** | {} |", cell.replace('|', "\\|"));
-            }
-            // Same rule as the dedup row: a report covering a fraction of the
-            // input list because the rest was already done must say so.
-            if let Some(r) = &m.resumed
-                && r.get("targets_skipped_completed")
-                    .and_then(|v| v.as_u64())
-                    .unwrap_or(0)
-                    > 0
-            {
-                let _ = writeln!(
-                    out,
-                    "| **Resumed** | {} target(s) skipped as already completed (state file: {}) |",
-                    r.get("targets_skipped_completed")
-                        .and_then(|v| v.as_u64())
-                        .unwrap_or(0),
-                    r.get("state_file")
-                        .and_then(|v| v.as_str())
-                        .unwrap_or("?")
-                        .replace('|', "\\|")
-                );
-            }
-            // Only rendered when true — a "Complete: yes" row on every clean
-            // report is noise, but its absence must never be what signals a
-            // truncated run, so the true case is spelled out loudly.
-            if m.incomplete {
-                let _ = writeln!(
-                    out,
-                    "| **Incomplete** | ⚠️ yes — at least one target was not fully tested (see Target Summary) |"
-                );
-            }
-            out.push('\n');
-
-            // Per-target summary table (includes status, findings_count, WAF when present)
-            if !m.target_summary.is_empty() {
-                out.push_str("### Target Summary\n\n");
-                out.push_str("| Target | Status | Findings | WAF |\n");
-                out.push_str("|--------|--------|----------|-----|\n");
-                for t in &m.target_summary {
-                    let tgt = t.get("target").and_then(|v| v.as_str()).unwrap_or("?");
-                    let st = t.get("status").and_then(|v| v.as_str()).unwrap_or("?");
-                    let fc = t
-                        .get("findings_count")
-                        .and_then(|v| v.as_u64())
-                        .unwrap_or(0);
-                    let status_cell = if let Some(ec) = t.get("error_code").and_then(|e| e.as_str())
-                    {
-                        format!("{} ({})", st, ec)
-                    } else {
-                        st.to_string()
-                    };
-                    let waf_str = if let Some(w) = t.get("waf") {
-                        // Real shape (from analysis.rs + render_results): "detected": [{ "type": "..", "confidence": N, ...}, ...]
-                        // plus optional "bypass". Support legacy test mock shape {detected: bool, name} too.
-                        if let Some(dets) = w.get("detected").and_then(|d| d.as_array()) {
-                            if !dets.is_empty() {
-                                dets[0]
-                                    .get("type")
-                                    .and_then(|ty| ty.as_str())
-                                    .unwrap_or("detected")
-                                    .to_string()
-                            } else {
-                                "none".to_string()
-                            }
-                        } else if w.get("detected").and_then(|d| d.as_bool()).unwrap_or(false) {
-                            w.get("name")
-                                .and_then(|n| n.as_str())
-                                .unwrap_or("detected")
-                                .to_string()
-                        } else {
-                            "none".to_string()
-                        }
-                    } else {
-                        "none".to_string()
-                    };
-                    let _ = writeln!(
-                        out,
-                        "| {} | {} | {} | {} |",
-                        tgt.replace('|', "\\|"),
-                        status_cell.replace('|', "\\|"),
-                        fc,
-                        waf_str.replace('|', "\\|")
-                    );
-                }
-                out.push('\n');
-            }
-        }
-
-        // Add summary
-        let v_count = results
-            .iter()
-            .filter(|r| r.result_type == FindingType::Verified)
-            .count();
-        let r_count = results
-            .iter()
-            .filter(|r| r.result_type == FindingType::Reflected)
-            .count();
-        out.push_str("## Summary\n\n");
-        let _ = writeln!(out, "- **Total Findings**: {}", results.len());
-        let _ = writeln!(out, "- **Vulnerabilities (V)**: {}", v_count);
-        let _ = write!(out, "- **Reflections (R)**: {}\n\n", r_count); // double newline intentional
-
-        // Add findings table
-        if !results.is_empty() {
-            out.push_str("## Findings\n\n");
-
-            for (idx, result) in results.iter().enumerate() {
-                let _ = write!(
-                    out,
-                    "### {}. {} - {} ({})\n\n", // double newline intentional
-                    idx + 1,
-                    if result.result_type == FindingType::Verified {
-                        "Vulnerability"
-                    } else {
-                        "Reflection"
-                    },
-                    result.param,
-                    result.inject_type
-                );
-
-                out.push_str("| Field | Value |\n");
-                out.push_str("|-------|-------|\n");
-                let _ = writeln!(out, "| **Type** | {} |", result.result_type);
-                let _ = writeln!(out, "| **Parameter** | `{}` |", result.param);
-                let _ = writeln!(out, "| **Method** | {} |", result.method);
-                let _ = writeln!(out, "| **Injection Type** | {} |", result.inject_type);
-                let _ = writeln!(
-                    out,
-                    "| **Detected By** | {} |",
-                    result.detection_method.as_str()
-                );
-                if let Some(grade) = result.confidence {
-                    let cell = if result.confidence_reason.is_empty() {
-                        grade.as_str().to_string()
-                    } else {
-                        format!(
-                            "{} ({})",
-                            grade.as_str(),
-                            result.confidence_reason.replace('|', "\\|")
-                        )
-                    };
-                    let _ = writeln!(out, "| **Confidence** | {} |", cell);
-                }
-                if let Some(is_new) = result.new_since_baseline {
-                    let _ = writeln!(
-                        out,
-                        "| **New** | {} |",
-                        if is_new { "yes" } else { "no (in baseline)" }
-                    );
-                }
-                let _ = writeln!(out, "| **Severity** | {} |", result.severity);
-                let _ = writeln!(out, "| **CWE** | {} |", result.cwe);
-                let _ = writeln!(out, "| **URL** | {} |", result.data);
-                let _ = writeln!(
-                    out,
-                    "| **Payload** | `{}` |",
-                    result.payload.replace('|', "\\|")
-                );
-
-                if !result.evidence.is_empty() {
-                    let _ = writeln!(
-                        out,
-                        "| **Evidence** | {} |",
-                        result.evidence.replace('|', "\\|")
-                    );
-                }
-
-                out.push('\n');
-
-                // Include request if requested
-                if include_request && let Some(req) = &result.request {
-                    out.push_str("**Request:**\n\n```http\n");
-                    out.push_str(req);
-                    out.push_str("\n```\n\n");
-                }
-
-                // Include response if requested
-                if include_response && let Some(resp) = &result.response {
-                    out.push_str("**Response:**\n\n```http\n");
-                    out.push_str(resp);
-                    out.push_str("\n```\n\n");
-                }
-
-                out.push_str("---\n\n");
-            }
-        }
-
-        out
-    }
-
-    /// Serialize a slice of Result into SARIF v2.1.0 format string.
-    /// SARIF (Static Analysis Results Interchange Format) is a standard format for static analysis tools.
-    ///
-    /// For backward compatibility (public API surface under `dalfox::scanning::result`),
-    /// the 3-argument form omits the scan metadata envelope. Use the `_with_meta` variant
-    /// to populate `run.properties` + `tool.driver.properties` (recommended for CI/code-scanning).
-    pub fn results_to_sarif(
-        results: &[Result],
-        include_request: bool,
-        include_response: bool,
-    ) -> String {
-        Self::results_to_sarif_with_meta(results, include_request, include_response, None)
-    }
-
-    /// Serialize ... with optional scan metadata (see `results_to_sarif`).
-    pub(crate) fn results_to_sarif_with_meta(
-        results: &[Result],
-        include_request: bool,
-        include_response: bool,
-        meta: Option<&ScanMetadata>,
-    ) -> String {
-        use serde_json::json;
-
-        // Convert severity to SARIF level
-        let severity_to_level = |severity: &str| -> &str {
-            match severity.to_lowercase().as_str() {
-                "high" | "critical" => "error",
-                "medium" => "warning",
-                "low" | "info" => "note",
-                _ => "warning",
-            }
-        };
-
-        // SARIF requires that a result's `ruleId` reference a rule defined in
-        // `driver.rules`, and that a present `ruleIndex` point at THAT rule.
-        // Findings carry different CWEs (XSS is CWE-79; outdated-library
-        // findings are CWE-1104), so a single hardcoded rule + `ruleIndex: 0`
-        // emitted an id with no matching rule and an inconsistent index, which
-        // GitHub code scanning rejects. Build the rule table from the distinct
-        // CWEs actually present, in first-seen order, and index each result
-        // into it.
-        let mut rule_ids: Vec<String> = Vec::new();
-        for r in results {
-            let id = sarif_rule_id(&r.cwe);
-            if !rule_ids.contains(&id) {
-                rule_ids.push(id);
-            }
-        }
-        // Never emit an empty `rules` array — keep the XSS rule as the default
-        // so an empty result set still describes the tool's primary rule.
-        if rule_ids.is_empty() {
-            rule_ids.push(sarif_rule_id("CWE-79"));
-        }
-        let rules: Vec<serde_json::Value> =
-            rule_ids.iter().map(|id| sarif_rule_for_id(id)).collect();
-
-        // Convert results to SARIF result objects
-        let sarif_results: Vec<serde_json::Value> = results
-            .iter()
-            .map(|r| {
-                // Build message with additional context
-                let mut message_parts = vec![r.message_str.clone()];
-                if !r.evidence.is_empty() {
-                    message_parts.push(format!("Evidence: {}", r.evidence));
-                }
-                if include_request && r.request.is_some() {
-                    message_parts.push("HTTP request included in properties".to_string());
-                }
-                if include_response && r.response.is_some() {
-                    message_parts.push("HTTP response included in properties".to_string());
-                }
-                let full_message = message_parts.join(". ");
-
-                // Build properties bag
-                let mut properties = json!({
-                    "type": r.result_type,
-                    "inject_type": r.inject_type,
-                    "method": r.method,
-                    "param": r.param,
-                    "payload": r.payload,
-                    "severity": r.severity,
-                    "detection_method": r.detection_method.as_str(),
-                });
-                if let Some(grade) = r.confidence {
-                    properties["confidence"] = json!(grade.as_str());
-                    if !r.confidence_reason.is_empty() {
-                        properties["confidence_reason"] = json!(r.confidence_reason);
-                    }
-                }
-
-                if let Some(is_new) = r.new_since_baseline {
-                    properties["new"] = json!(is_new);
-                }
-
-                if include_request && let Some(req) = &r.request {
-                    properties["request"] = json!(req);
-                }
-                if include_response && let Some(resp) = &r.response {
-                    properties["response"] = json!(resp);
-                }
-
-                // Stable, vulnerability-identity fingerprint so SARIF
-                // consumers (e.g. GitHub code scanning) can dedupe the
-                // same finding across rescans. Previously this was the
-                // catalog `message_id`, which is hardcoded per finding
-                // type (e.g. 606 for every reflected XSS) and therefore
-                // useless for dedup.
-                let stable_fp = crate::utils::stable_finding_fingerprint(
-                    &r.data,
-                    &r.param,
-                    &r.inject_type,
-                    &r.cwe,
-                );
-                let rule_id = sarif_rule_id(&r.cwe);
-                let rule_index = rule_ids.iter().position(|id| *id == rule_id).unwrap_or(0);
-                json!({
-                    "ruleId": rule_id,
-                    "ruleIndex": rule_index,
-                    "level": severity_to_level(&r.severity),
-                    "message": {
-                        "text": full_message
-                    },
-                    "locations": [{
-                        "physicalLocation": {
-                            "artifactLocation": {
-                                "uri": r.data.clone()
-                            },
-                            "region": {
-                                "snippet": {
-                                    "text": r.payload.clone()
-                                }
-                            }
-                        }
-                    }],
-                    "partialFingerprints": {
-                        // SARIF spec: keys are arbitrary identifiers, values
-                        // are stable hashes. v1 versions the scheme so we can
-                        // evolve the input tuple later without re-mapping
-                        // historical findings.
-                        "vulnIdentity/v1": stable_fp,
-                        // Preserve the catalog id under a clearly non-
-                        // fingerprint name — useful for human triage but
-                        // not used by consumers for dedup.
-                        "dalfoxMessageId": r.message_id.to_string(),
-                    },
-                    "properties": properties
-                })
-            })
-            .collect();
-
-        // Build driver, optionally with scan meta under its properties (per issue #1093)
-        let mut driver = json!({
-            "name": "Dalfox",
-            "informationUri": "https://github.com/hahwul/dalfox",
-            "version": env!("CARGO_PKG_VERSION"),
-            "rules": rules,
-        });
-        if let Some(m) = meta {
-            driver["properties"] = Self::make_scan_meta_value(m);
-        }
-
-        // Build run object, optionally with scan meta under run.properties
-        let mut run = json!({
-            "tool": {
-                "driver": driver
-            },
-            "results": sarif_results
-        });
-        if let Some(m) = meta {
-            run["properties"] = Self::make_scan_meta_value(m);
-        }
-
-        // Build SARIF document
-        let sarif = json!({
-            "version": "2.1.0",
-            "$schema": "https://raw.githubusercontent.com/oasis-tcs/sarif-spec/master/Schemata/sarif-schema-2.1.0.json",
-            "runs": [run]
-        });
-
-        serde_json::to_string_pretty(&sarif).unwrap_or_else(|_| "{}".to_string())
-    }
 }
 
-/// SARIF `ruleId` / `driver.rules[].id` for a finding's CWE. Empty or unknown
-/// CWEs fall back to the XSS rule so the id is always a well-formed
-/// `dalfox/cwe-<n>` rather than a bare `dalfox/`.
-fn sarif_rule_id(cwe: &str) -> String {
-    let trimmed = cwe.trim();
-    if trimmed.is_empty() {
-        return "dalfox/cwe-79".to_string();
-    }
-    format!("dalfox/{}", trimmed.to_ascii_lowercase())
-}
-
-/// Build a `driver.rules[]` entry for a `dalfox/cwe-<n>` rule id. CWE-79 (the
-/// scanner's primary finding class) keeps its rich description; other CWEs get
-/// a minimal-but-valid rule so every emitted `ruleId` resolves to a defined
-/// rule (SARIF requirement for GitHub code scanning ingestion).
-fn sarif_rule_for_id(id: &str) -> serde_json::Value {
-    use serde_json::json;
-    match id {
-        "dalfox/cwe-1104" => json!({
-            "id": id,
-            "name": "UseOfUnmaintainedThirdPartyComponents",
-            "shortDescription": {
-                "text": "Use of a component with known vulnerabilities (CWE-1104)"
-            },
-            "fullDescription": {
-                "text": "The application loads a third-party JavaScript component whose version is known to be outdated or vulnerable."
-            },
-            "help": {
-                "text": "Upgrade the flagged component to a maintained, non-vulnerable release and track dependency advisories."
-            },
-            "defaultConfiguration": { "level": "warning" },
-            "properties": { "tags": ["security", "dependencies"], "precision": "high" }
-        }),
-        "dalfox/cwe-79" => json!({
-            "id": id,
-            "name": "CrossSiteScripting",
-            "shortDescription": {
-                "text": "Cross-site Scripting (XSS)"
-            },
-            "fullDescription": {
-                "text": "The application reflects user input in HTML responses without proper encoding, allowing attackers to inject malicious scripts."
-            },
-            "help": {
-                "text": "Ensure all user input is properly encoded before being rendered in HTML context. Use context-aware output encoding based on where the data is placed (HTML body, attributes, JavaScript, CSS, or URL)."
-            },
-            "defaultConfiguration": {
-                "level": "error"
-            },
-            "properties": {
-                "tags": ["security", "xss", "injection"],
-                "precision": "high"
-            }
-        }),
-        // Any other CWE: emit a valid, generic rule so the id resolves.
-        other => json!({
-            "id": other,
-            "name": "SecurityFinding",
-            "shortDescription": { "text": format!("Security finding ({})", other) },
-            "defaultConfiguration": { "level": "warning" },
-            "properties": { "tags": ["security"] }
-        }),
-    }
-}
+mod format_json;
+mod format_markdown;
+mod format_sarif;
+mod format_toml;
 
 #[cfg(test)]
 mod tests;
